+++ apps/erlangbridge/src/lsp_syntax.erl
@@ -177,6 +177,7 @@ get_standard_include_paths() ->
     RootDir = gen_lsp_config_server:root(),
     [
         filename:join([RootDir, "_build", "default", "lib"]),
+        filename:join([RootDir, "_build", "default", "plugins"]),
         filename:join([RootDir, "apps"]),
         filename:join([RootDir, "lib"])
     ].
@@ -255,8 +256,19 @@ get_include_paths_from_rebar_config(File) ->
 			   _ -> 
                    []
 			 end,
+        Deps = [],
+        %TODO: if include of each rebar dependency should be add to include paths, uncomment below 
+    %   Deps = case Consult of
+    %       {ok, DepsTerms} ->
+    %         Profiles = proplists:get_value(profiles, DepsTerms, []),
+    %         ErlDeps = lists:flatmap(fun({_,Opts})-> proplists:get_value(deps,Opts) end, Profiles),
+    %         RootDir = gen_lsp_config_server:root(),
+    %         [filename:join([RootDir, "_build", "default", "plugins", X, "include"]) || X <- ErlDeps ];
+    %       _ -> 
+    %         []
+    %       end,
 	  DefaultPaths = [filename:dirname(RebarConfig), filename:join([filename:dirname(RebarConfig), "include"])],
-	  ErlOptsPaths ++ DefaultPaths
+	  ErlOptsPaths ++ Deps ++ DefaultPaths
     end.
 
 find_rebar_config(Dir) ->
@@ -301,7 +313,7 @@ lint(FileSyntaxTree, File) ->
       {ok, [Warnings]} ->
 	  #{parse_result => true,
 	    errors_warnings =>
-		extract_error_or_warning(<<"warning">>, Warnings)};
+		extract_error_or_warning(<<"warning">>, filter_unused_functions(Warnings))};
       % errors, no warnings
       {error, [Errors], []} ->
 	  #{parse_result => true,
@@ -312,14 +324,34 @@ lint(FileSyntaxTree, File) ->
 	  #{parse_result => true,
 	    errors_warnings =>
 		extract_error_or_warning(<<"error">>, Errors) ++
-		  extract_error_or_warning(<<"warning">>, Warnings)};
+		  extract_error_or_warning(<<"warning">>, filter_unused_functions(Warnings))};
       {error, [], [Warnings]} ->
 	  #{parse_result => true,
 	    errors_warnings =>
-		extract_error_or_warning(<<"warning">>, Warnings)};
+		extract_error_or_warning(<<"warning">>, filter_unused_functions(Warnings))};
       _Any ->
 	  #{parse_result => false, error_message => <<"lint error">>}
     end.
+
+filter_unused_functions({_, []}) ->
+    [];
+filter_unused_functions({File, Warnings}) ->
+    %Filter unused function that ends with "_test", to avoid unwanted warnings in unit tests modules
+    %%TODO: if more than one filter to exclude, it should be configurable
+    Result = {
+        File, 
+        lists:filter(fun (X) ->
+                case X of
+                    {_, _, {unused_function, {FuncName,_}}} -> 
+                        FindResult = string:find(atom_to_list(FuncName), "_test", trailing) =/= "_test",
+                        %gen_lsp_server:lsp_log("FuncName:~p, ~p",[FuncName, FindResult]),
+                        FindResult;
+                    _ -> true 
+                end 
+            end,
+            Warnings)
+    }, 
+    Result.
 
 extract_error_or_warning(_Type, {_, []}) ->
     [];