+++ lib/erlangDebug.ts
@@ -1,4 +1,4 @@
-import { DebugSession } from 'vscode-debugadapter';
 import { ErlangDebugSession } from './erlangDebugSession';
 
-DebugSession.run(ErlangDebugSession);    
+ErlangDebugSession.run(ErlangDebugSession);    
+